Document zones and simplify zone allocation API
Fixes #779
