+++ javalib/src/main/scala/java/util/regex/cre2h.scala
@@ -29,7 +29,7 @@ object cre2h {
   }
 
   def toRE2String(str: String, restr: Ptr[cre2.string_t])(
-      implicit a: Alloc): Unit = {
+      implicit z: Zone): Unit = {
     restr.data = toCString(str)
     restr.length = str.length
   }
