+++ unit-tests/src/main/scala/scala/scalanative/native/CStringSuite.scala
@@ -4,7 +4,6 @@ import stdio._
 import string._
 
 object CStringSuite extends tests.Suite {
-  implicit val alloc = Alloc.system
 
   test("fromCString") {
     val cstrFrom = c"1234"
@@ -18,15 +17,17 @@ object CStringSuite extends tests.Suite {
   }
 
   test("toCString") {
-    val szFrom = "abcde"
-    val cstrTo = toCString(szFrom)
+    Zone { implicit z =>
+      val szFrom = "abcde"
+      val cstrTo = toCString(szFrom)
 
-    assert(strlen(cstrTo) == 5)
-    assert(cstrTo(0) == 'a'.toByte)
-    assert(cstrTo(1) == 'b'.toByte)
-    assert(cstrTo(2) == 'c'.toByte)
-    assert(cstrTo(3) == 'd'.toByte)
-    assert(cstrTo(4) == 'e'.toByte)
-    assert(cstrTo(5) == '\0'.toByte)
+      assert(strlen(cstrTo) == 5)
+      assert(cstrTo(0) == 'a'.toByte)
+      assert(cstrTo(1) == 'b'.toByte)
+      assert(cstrTo(2) == 'c'.toByte)
+      assert(cstrTo(3) == 'd'.toByte)
+      assert(cstrTo(4) == 'e'.toByte)
+      assert(cstrTo(5) == '\0'.toByte)
+    }
   }
 }
